Temporarily fix CI by skip internal message check
